% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editAddin.R
\name{editAddin}
\alias{editAddin}
\title{Edit a Data Frame.}
\usage{
editAddin()
}
\description{
Interactively edit a \code{data.frame} or \code{data.table}. The resulting
code will be emitted as a call to reload the data from a temp RDS file.
}
\details{
This addin can be used to interactively edit.
The intended way to use this is as follows:

1. Highlight a symbol naming a \code{data.frame} or \code{data.table}
   in your R session, e.g. \code{mtcars}.
2. Execute this addin, to interactively edit it.

When you're done, the code performing this operation will be emitted
at the cursor position.

This function borrows heavily from \href{rstudio/addinexamples/subsetAddin}{https://github.com/rstudio/addinexamples/blob/master/R/subsetAddin.R}
}
