% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_cell}
\alias{hot_cell}
\title{Configure single cell.  See
\href{http://handsontable.com}{Handsontable.js} for details.}
\usage{
hot_cell(hot, row, col, comment = NULL)
}
\arguments{
\item{hot}{rhandsontable object}

\item{row}{numeric row index}

\item{col}{numeric column index}

\item{comment}{character comment to add to cell}
}
\description{
Configure single cell.  See
\href{http://handsontable.com}{Handsontable.js} for details.
}

