% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_col}
\alias{hot_col}
\title{Configure single column.  See
\href{http://handsontable.com}{Handsontable.js} for details.}
\usage{
hot_col(hot, col, type, format = NULL, source = NULL, strict = NULL,
  allowInvalcol = NULL, readOnly = FALSE, validator = NULL,
  halign = NULL, valign = NULL, renderer = NULL)
}
\arguments{
\item{hot}{rhandsontable object}

\item{col}{}

\item{type}{}

\item{format}{}

\item{source}{}

\item{strict}{}

\item{allowInvalcol}{}

\item{readOnly}{}

\item{validator}{regex expression or JS()}

\item{halign}{htLeft, htCenter, htRight, htJustify}

\item{valign}{htTop, htMiddle, htBottom}

\item{renderer}{}
}
\description{
Configure single column.  See
\href{http://handsontable.com}{Handsontable.js} for details.
}

