% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_cols}
\alias{hot_cols}
\title{Configure column parameters.  See
\href{http://handsontable.com}{Handsontable.js} for details.}
\usage{
hot_cols(hot, columns = NULL, colWidths = NULL, columnSorting = FALSE,
  manualColumnMove = FALSE, manualColumnResize = FALSE,
  fixedColumnsLeft = NULL, halign = NULL, valign = NULL, ...)
}
\arguments{
\item{hot}{rhandsontable object}

\item{columns}{}

\item{colWidths}{scalar, vector or JS()}

\item{columnSorting}{}

\item{manualColumnMove}{}

\item{manualColumnResize}{}

\item{fixedColumnsLeft}{}

\item{halign}{htLeft, htCenter, htRight, htJustify}

\item{valign}{htTop, htMiddle, htBottom}

\item{...}{passed to hot_col}
}
\description{
Configure column parameters.  See
\href{http://handsontable.com}{Handsontable.js} for details.
}

