% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_condformat}
\alias{hot_condformat}
\title{Add conditional formatting to column.  See
\href{http://handsontable.com/demo/conditional.html}{Conditional formatting}
for details.}
\usage{
hot_condformat(hot, col, renderer)
}
\arguments{
\item{hot}{rhandsontable object}

\item{col}{}

\item{renderer}{}
}
\description{
Add conditional formatting to column.  See
\href{http://handsontable.com/demo/conditional.html}{Conditional formatting}
for details.
}

