% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_context_menu}
\alias{hot_context_menu}
\title{Configure the options for the right-click context menu}
\usage{
hot_context_menu(hot, allowRowEdit = TRUE, allowColEdit = TRUE,
  allowReadOnly = FALSE, allowComments = TRUE, allowCustomBorders = FALSE,
  exportToCsv = FALSE, csvFileName = "download.csv", ...)
}
\arguments{
\item{hot}{rhandsontable object}

\item{allowRowEdit}{logical enabling row editing}

\item{allowColEdit}{logical enabling column editing ***ADD MORE}

\item{allowReadOnly}{logical enabling read-only toggle}

\item{allowComments}{logical enabling comments}

\item{allowCustomBorders}{logical enabling custom borders}

\item{exportToCsv}{logical adding a context menu option to export the table
data to a csv file}

\item{csvFileName}{character csv file name}

\item{...}{ignored}
}
\description{
Configure the options for the right-click context menu
}

