% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_heatmap}
\alias{hot_heatmap}
\title{Add heatmap to table.  See
\href{http://handsontable.com/demo/heatmaps.html}{Heatmaps for values in a column}
for details.}
\usage{
hot_heatmap(hot, cols, color_scale, renderer = NULL)
}
\arguments{
\item{hot}{rhandsontable object}

\item{cols}{}

\item{color_scale}{}

\item{renderer}{}
}
\description{
Add heatmap to table.  See
\href{http://handsontable.com/demo/heatmaps.html}{Heatmaps for values in a column}
for details.
}

