% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_to_r}
\alias{hot_to_r}
\title{Handsontable widget}
\usage{
hot_to_r(...)
}
\arguments{
\item{...}{passed to \code{rhandsontable:::toR}}
}
\description{
Convert handsontable data to R object. Can be used in a \code{shiny} app
 to convert the input json to an R dataset.
}
\seealso{
\code{\link{rHandsontableOutput}}
}

