% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_validate_character}
\alias{hot_validate_character}
\title{Add numeric validation to a column}
\usage{
hot_validate_character(hot, col, choices, allowInvalid = FALSE)
}
\arguments{
\item{hot}{rhandsontable object}

\item{col}{}

\item{choices}{}

\item{allowInvalid}{}
}
\description{
Add numeric validation to a column
}

