% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_validate_numeric}
\alias{hot_validate_numeric}
\title{Add numeric validation to a column}
\usage{
hot_validate_numeric(hot, col, min = NULL, max = NULL, choices = NULL,
  exclude = NULL, allowInvalid = FALSE)
}
\arguments{
\item{hot}{rhandsontable object}

\item{col}{}

\item{min}{}

\item{max}{}

\item{choices}{}

\item{exclude}{}

\item{allowInvalid}{}
}
\description{
Add numeric validation to a column
}

