% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhandsontable-package.R
\name{rhandsontable-exports}
\alias{\%>\%}
\alias{rhandsontable-exports}
\title{rhandsontable exported operators}
\description{
The following functions are imported and then re-exported
from the rhandsontable package to enable use of the magrittr
pipe operator with no additional library calls
}

