% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{rhandsontable}
\alias{rhandsontable}
\title{rhandsontable}
\usage{
rhandsontable(data, rownames = NULL, width = NULL, height = NULL)
}
\description{
htmlwidgets implementation of handsontable.js
}

