% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{rHandsontableOutput}
\alias{rHandsontableOutput}
\title{Widget output function for use in Shiny}
\usage{
rHandsontableOutput(outputId, width = NA, height = NA)
}
\description{
Widget output function for use in Shiny
}

