% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{set_data}
\alias{set_data}
\title{Handsontable widget}
\usage{
set_data(id, row, col, val, session, zero_indexed = F)
}
\arguments{
\item{id}{The id of the table to interact with.}

\item{row}{Integer vector of row indexes.}

\item{col}{Integer vector the column indexes.}

\item{val}{Vector of values to set at each row-col pair.}

\item{session}{The session that is associated with your shiny server function. The table is only interactive when used in shiny so we only use set_data when the table is in shiny.}

\item{zero_indexed}{Default FALSE. Set to TRUE if you are supplying row and col indexes that are already 0-based.}
}
\description{
Set data inside a Handsontable instance without recreating the widget. Send the new values as a vector of rows, a vector of columns, and a vector of values. If different length vectors are supplied then the shorter ones are recycled to match the length of the longest.
}
