% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_col}
\alias{hot_col}
\title{Configure single column.  See
\href{http://handsontable.com}{Handsontable.js} for details.}
\usage{
hot_col(hot, col, type = NULL, format = NULL, source = NULL,
  strict = NULL, readOnly = NULL, validator = NULL, allowInvalid = NULL,
  halign = NULL, valign = NULL, renderer = NULL)
}
\arguments{
\item{hot}{rhandsontable object}

\item{col}{numeric column index}

\item{type}{character specify the data type. Options include:
numeric, date, checkbox, select, dropdown, autocomplete, password,
and handsontable (not implemented yet)}

\item{format}{characer specifying column format. See Cell Types at
\href{http://handsontable.com}{Handsontable.js} for the formatting
options for each data type}

\item{source}{a vector of choices for select, dropdown and autocomplete
column types}

\item{strict}{logical specifying whether values not in the \code{source}
vector will be accepted}

\item{readOnly}{logical making the table read-only}

\item{validator}{character defining a Javascript function to be used
to validate user input. See \code{hot_validate_numeric} and
\code{hot_validate_character} for pre-build validators.}

\item{allowInvalid}{logical specifying whether invalid data will be
accepted. Invalid data cells will be color red.}

\item{halign}{character defining the horizontal alignment. Possible
values are htLeft, htCenter, htRight and htJustify}

\item{valign}{character defining the vertical alignment. Possible
values are htTop, htMiddle, htBottom}

\item{renderer}{character defining a Javascript function to be used
to format column cells. Can be used to implement conditional formatting.}
}
\description{
Configure single column.  See
\href{http://handsontable.com}{Handsontable.js} for details.
}

