% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_col}
\alias{hot_col}
\title{Configure single column.  See
\href{http://handsontable.com}{Handsontable.js} for details.}
\usage{
hot_col(hot, col, type = NULL, format = NULL, source = NULL,
  strict = NULL, readOnly = NULL, validator = NULL, allowInvalid = NULL,
  halign = NULL, valign = NULL, renderer = NULL)
}
\arguments{
\item{hot}{rhandsontable object}

\item{col}{}

\item{type}{}

\item{format}{}

\item{source}{}

\item{strict}{}

\item{readOnly}{}

\item{validator}{}

\item{allowInvalid}{}

\item{halign}{htLeft, htCenter, htRight, htJustify}

\item{valign}{htTop, htMiddle, htBottom}

\item{renderer}{}
}
\description{
Configure single column.  See
\href{http://handsontable.com}{Handsontable.js} for details.
}

