% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_rows}
\alias{hot_rows}
\title{Configure rows.  See
\href{http://handsontable.com}{Handsontable.js} for details.}
\usage{
hot_rows(hot, rowHeights = NULL, fixedRowsTop = NULL)
}
\arguments{
\item{hot}{rhandsontable object}

\item{rowHeights}{scalar, vector or JS()}

\item{fixedRowsTop}{}
}
\description{
Configure rows.  See
\href{http://handsontable.com}{Handsontable.js} for details.
}

