% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{hot_table}
\alias{hot_table}
\title{Configure table.  See
\href{http://handsontable.com}{Handsontable.js} for details.}
\usage{
hot_table(hot, customBorders = NULL, contextMenu = TRUE, groups = NULL,
  highlightRow = FALSE, highlightCol = FALSE, wordWrap = TRUE)
}
\arguments{
\item{hot}{rhandsontable object}

\item{customBorders}{json See
\href{http://handsontable.com/demo/custom_borders.html}{Custom borders} for details.}

\item{contextMenu}{}

\item{groups}{json See
\href{http://handsontable.com/demo/grouping.html}{Grouping & ungrouping of rows and columns} for details.}

\item{highlightRow}{}

\item{highlightCol}{}

\item{wordWrap}{}
}
\description{
Configure table.  See
\href{http://handsontable.com}{Handsontable.js} for details.
}

