% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rHandsontable.R
\name{renderRHandsontable}
\alias{renderRHandsontable}
\title{Widget render function for use in Shiny}
\usage{
renderRHandsontable(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

