% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{renderRHandsontable}
\alias{renderRHandsontable}
\title{Handsontable widget}
\usage{
renderRHandsontable(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates an rhandsontable.}

\item{env}{the environment in which to evaluate \code{expr}.}

\item{quoted}{is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Shiny bindings for rhandsontable
}
\seealso{
\code{\link{rHandsontableOutput}}, \code{\link{hot_to_r}}
}

