% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhandsontable-package.R
\docType{package}
\name{rhandsontable-package}
\alias{rhandsontable-package}
\title{rhandsontable}
\description{
R interface for creating tables using Handsontable, url{http://http://handsontable.com/}
}
\details{
For full documentation on the package, visit \url{http://jrowen.github.io/rhandsontable/}
}

