% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable-package.R, R/rhandsontable.R
\docType{package}
\name{rhandsontable}
\alias{rhandsontable}
\alias{rhandsontable-package}
\title{An \code{htmlwidget} interface to the
\href{http://http://handsontable.com//}{Handsontable.js} library}
\usage{
rhandsontable(data, colHeaders = NULL, rowHeaders = NULL, useTypes = TRUE,
  readOnly = NULL, contextMenu = TRUE, width = NULL, height = NULL)
}
\arguments{
\item{data}{}

\item{colHeaders}{}

\item{rowHeaders}{}

\item{useTypes}{}

\item{readOnly}{}

\item{contextMenu}{}

\item{width}{}

\item{height}{}
}
\description{
An \code{htmlwidget} interface to the
\href{http://http://handsontable.com//}{Handsontable.js} library

See \href{http://handsontable.com}{Handsontable.js} for details.
}
\author{
Jonathan Owen
}

