% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhandsontable-package.R, R/rhandsontable.R
\docType{package}
\name{rhandsontable}
\alias{rhandsontable}
\alias{rhandsontable-package}
\title{An \code{htmlwidget} interface to the
\href{http://http://handsontable.com//}{Handsontable.js} library}
\usage{
rhandsontable(data, colHeaders, rowHeaders, useTypes = TRUE,
  readOnly = NULL, contextMenu = TRUE, width = NULL, height = NULL)
}
\arguments{
\item{data}{a \code{data.table}, \code{data.frame} or \code{matrix}}

\item{colHeaders}{a vector of column names. If missing \code{colnames}
will be used. Setting to \code{NULL} will omit.}

\item{rowHeaders}{a vector of row names. If missing \code{rownames}
will be used. Setting to \code{NULL} will omit.}

\item{useTypes}{logical specifying whether column classes should be mapped to
equivalent Javascript types}

\item{readOnly}{logical specifying whether the table is editable}

\item{contextMenu}{passed to \code{hot_table}}

\item{width}{numeric table width}

\item{height}{numeric table height}
}
\description{
An \code{htmlwidget} interface to the
\href{http://http://handsontable.com//}{Handsontable.js} library

See \href{http://handsontable.com}{Handsontable.js} for details.
}
\author{
Jonathan Owen
}

