% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rHandsontable.R
\name{rHandsontableOutput}
\alias{rHandsontableOutput}
\title{Widget output function for use in Shiny}
\usage{
rHandsontableOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

