% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{rHandsontableOutput}
\alias{rHandsontableOutput}
\title{Handsontable widget}
\usage{
rHandsontableOutput(outputId, width = NA, height = NA)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width,height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}
}
\description{
Shiny bindings for rhandsontable
}

