% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhandsontable.R
\name{rHandsontableOutput}
\alias{rHandsontableOutput}
\title{Handsontable widget}
\usage{
rHandsontableOutput(outputId, width = "100\%", height = "100\%")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{must be a valid CSS unit in pixels
or a number, which will be coerced to a string and have \code{"px"} appended.}
}
\description{
Shiny bindings for rhandsontable
}
\seealso{
\code{\link{renderRHandsontable}}
}

